restructure to standard python

+++ deepspeech/helper_routines.py
@@ -5,7 +5,6 @@ and apply weight decay.
 import re
 import tensorflow as tf
 
-
 # If a model is trained with multiple GPUs, prefix all Op names with tower_name
 # to differentiate the operations. Note that this prefix is removed from the
 # names of the summaries when visualizing a model.
@@ -28,7 +27,11 @@ def _activation_summary(act):
     tf.summary.scalar(tensor_name + '/sparsity', tf.nn.zero_fraction(act))
 
 
-def _variable_on_cpu(name, shape, initializer = None, use_fp16 = False, trainable = True):
+def _variable_on_cpu(name,
+                     shape,
+                     initializer=None,
+                     use_fp16=False,
+                     trainable=True):
     """Helper to create a Variable stored on cpu memory.
 
     Args:
@@ -41,8 +44,12 @@ def _variable_on_cpu(name, shape, initializer = None, use_fp16 = False, trainabl
     """
     with tf.device('/cpu'):
         dtype = tf.float16 if use_fp16 else tf.float32
-        var = tf.get_variable(name, shape,
-                              initializer = initializer, dtype = dtype, trainable = trainable)
+        var = tf.get_variable(
+            name,
+            shape,
+            initializer=initializer,
+            dtype=dtype,
+            trainable=trainable)
     return var
 
 
@@ -62,17 +69,16 @@ def _variable_with_weight_decay(name, shape, wd_value, use_fp16):
       Variable Tensor
     """
     dtype = tf.float16 if use_fp16 else tf.float32
-    var = _variable_on_cpu(
-        name,
-        shape,
-        tf.contrib.layers.variance_scaling_initializer(factor = 2.0,
-                                                       mode = 'FAN_IN',
-                                                       uniform = False,
-                                                       seed = None,
-                                                       dtype = dtype), use_fp16)
+    var = _variable_on_cpu(name, shape,
+                           tf.contrib.layers.variance_scaling_initializer(
+                               factor=2.0,
+                               mode='FAN_IN',
+                               uniform=False,
+                               seed=None,
+                               dtype=dtype), use_fp16)
     if wd_value is not None:
-        weight_decay = tf.cast(tf.mul(tf.nn.l2_loss(var),
-                                      wd_value, name = 'weight_loss'),
-                               tf.float32)  # CTC loss is in float32
+        weight_decay = tf.cast(
+            tf.mul(tf.nn.l2_loss(var), wd_value, name='weight_loss'),
+            tf.float32)  # CTC loss is in float32
         tf.add_to_collection('losses', weight_decay)
     return var
